Extracted livereload & removed need for rice boxes
