+++ README.md
@@ -149,7 +149,6 @@ Haste is in its early days at the moment and I'm no golang pro so bugs are highl
 This software includes works from the following great open source projects:
 
 * [Golang](https://github.com/golang/go) - [License](https://github.com/golang/go/blob/master/LICENSE)
-* [Go Rice](https://github.com/GeertJohan/go.rice) - [License](https://github.com/GeertJohan/go.rice/blob/master/LICENSE)
 * [Go Color](https://github.com/fatih/color) - [License](https://github.com/fatih/color/blob/master/LICENSE.md)
 * [Go fsnotify](https://github.com/howeyc/fsnotify) - [License](https://github.com/howeyc/fsnotify/blob/master/LICENSE)
 * [Livereload](https://github.com/livereload/livereload-js) - [License](https://github.com/livereload/livereload-js/blob/master/LICENSE)
