+++ server/livereload.go
@@ -1,0 +1,59 @@
+package server
+
+import (
+	"golang.org/x/net/websocket"
+	"io"
+)
+
+type livereloadResponse struct {
+	Command string `json:"command"`
+}
+
+type livereloadHello struct {
+	Command    string   `json:"command"`
+	Protocols  []string `json:"protocols"`
+	ServerName string   `json:"serverName"`
+}
+
+type livereloadChange struct {
+	Command string `json:"command"`
+	Path    string `json:"path"`
+	LiveCSS bool   `json:"liveCSS"`
+}
+
+func (s *Server) getLivereloadWsHandler() func(ws *websocket.Conn) {
+	return func(ws *websocket.Conn) {
+
+		s.sockets = append(s.sockets, ws)
+
+		for {
+			// websocket.Message.Send(ws, "Hello, Client!")
+			wsData := new(livereloadResponse)
+			err := websocket.JSON.Receive(ws, &wsData)
+			if err != nil && err != io.EOF {
+				check(err)
+				return
+			} else if err == io.EOF {
+				return
+			}
+
+			if wsData.Command == "hello" {
+				response := livereloadHello{
+					Command: "hello",
+					Protocols: []string{
+						"http://livereload.com/protocols/connection-check-1",
+						"http://livereload.com/protocols/official-7",
+						"http://livereload.com/protocols/official-8",
+						"http://livereload.com/protocols/official-9",
+						"http://livereload.com/protocols/2.x-origin-version-negotiation",
+					},
+					ServerName: "Webby",
+				}
+				websocket.JSON.Send(ws, response)
+			}
+
+		}
+
+	}
+
+}
